+++ src/app/core/interfaces/condition.ts
