+++ src/app/core/interfaces/location-stamp.ts
