+++ src/app/core/interfaces/record.ts
