+++ src/app/core/interfaces/snapshot.ts
