+++ src/app/core/services/snapshot.service.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Plugins, GeolocationPosition } from '@capacitor/core';
 import { GeolocationService } from './geolocation.service';
-import { LocationStamp } from '../../interfaces/location-stamp';
-import { Snapshot } from '../../interfaces/snapshot';
+import { LocationStamp } from '../interfaces/location-stamp';
+import { Snapshot } from '../interfaces/snapshot';
 
 const { Geolocation } = Plugins;
